Fix test_get_dependents_ordered() and rename to test_get_dependents_is_deterministic()
The pytest test_get_dependents_ordered() tested that the
outcome of list(OrderedDict) != list(set(OrderedDict)).
This is not a guaranteed behaviour of the python set() and
list() functions, although it often will be the case.
The actual intention of the test function was to test
whether pdf.get_dependents() would return the same
OrderedDict on each subsequent call, i.e. whether the
function is deterministic. The test function has
been rewritten to reflect that intent.